copy java files to project template on build and clean them afterwards
