--- build-artifacts/project-template-gradle/src/main/assets/internal/livesync.js
@@ -1,3 +1,0 @@
-if (global.__onLiveSync) {
-	global.__onLiveSync();
-}
