--- build-artifacts/project-template-gradle/src/main/java/com/tns/internal/Plugin.java
@@ -1,5 +1,0 @@
-package com.tns.internal;
-
-public interface Plugin {
-    boolean execute(android.content.Context context) throws Exception;
-}
