+++ test-app/app/src/main/assets/internal/ts_helpers.js
@@ -1,6 +1,9 @@
 (function() {
 	var __extends_ts = function (d, b) {
-	    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+		if (!b.extend) {
+			for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+		}
+
 	    function __() { this.constructor = d; }
 	    __.prototype = b.prototype;
 	    d.prototype = new __();
