+++ test-app/app/src/main/res/layout/error_activity.xml
@@ -11,8 +11,7 @@
         android:id="@+id/toolbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:minHeight="?attr/actionBarSize"
-        android:layout_alignBottom="@+id/tabLayout"/>
+        android:minHeight="?attr/actionBarSize"/>
 
     <android.support.v4.view.ViewPager
         android:id="@+id/pager"
