+++ test-app/app/src/main/res/layout/exception_tab.xml
@@ -11,7 +11,6 @@
 
     <Button
         android:id="@+id/btnCopyException"
-        style="@style/Widget.AppCompat.Button.Colored"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
